+++ src/components/main/main.jsx
@@ -1,0 +1,158 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+const Main = (props) => {
+  return (
+    <React.Fragment>
+      <div style={{display: `none`}}>
+        <svg xmlns="http://www.w3.org/2000/svg"><symbol id="icon-arrow-select" viewBox="0 0 7 4"><path fillRule="evenodd" clipRule="evenodd" d="M0 0l3.5 2.813L7 0v1.084L3.5 4 0 1.084V0z"></path></symbol><symbol id="icon-bookmark" viewBox="0 0 17 18"><path d="M3.993 2.185l.017-.092V2c0-.554.449-1 .99-1h10c.522 0 .957.41.997.923l-2.736 14.59-4.814-2.407-.39-.195-.408.153L1.31 16.44 3.993 2.185z"></path></symbol><symbol id="icon-star" viewBox="0 0 13 12"><path fillRule="evenodd" clipRule="evenodd" d="M6.5 9.644L10.517 12 9.451 7.56 13 4.573l-4.674-.386L6.5 0 4.673 4.187 0 4.573 3.549 7.56 2.483 12 6.5 9.644z"></path></symbol></svg>
+      </div>
+
+      <div className="page page--gray page--main">
+        <header className="header">
+          <div className="container">
+            <div className="header__wrapper">
+              <div className="header__left">
+                <a className="header__logo-link header__logo-link--active">
+                  <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
+                </a>
+              </div>
+              <nav className="header__nav">
+                <ul className="header__nav-list">
+                  <li className="header__nav-item user">
+                    <a className="header__nav-link header__nav-link--profile" href="#">
+                      <div className="header__avatar-wrapper user__avatar-wrapper">
+                      </div>
+                      <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
+                    </a>
+                  </li>
+                </ul>
+              </nav>
+            </div>
+          </div>
+        </header>
+
+        <main className="page__main page__main--index">
+          <h1 className="visually-hidden">Cities</h1>
+          <div className="tabs">
+            <section className="locations container">
+              <ul className="locations__list tabs__list">
+                <li className="locations__item">
+                  <a className="locations__item-link tabs__item" href="#">
+                    <span>Paris</span>
+                  </a>
+                </li>
+                <li className="locations__item">
+                  <a className="locations__item-link tabs__item" href="#">
+                    <span>Cologne</span>
+                  </a>
+                </li>
+                <li className="locations__item">
+                  <a className="locations__item-link tabs__item" href="#">
+                    <span>Brussels</span>
+                  </a>
+                </li>
+                <li className="locations__item">
+                  <a className="locations__item-link tabs__item tabs__item--active">
+                    <span>Amsterdam</span>
+                  </a>
+                </li>
+                <li className="locations__item">
+                  <a className="locations__item-link tabs__item" href="#">
+                    <span>Hamburg</span>
+                  </a>
+                </li>
+                <li className="locations__item">
+                  <a className="locations__item-link tabs__item" href="#">
+                    <span>Dusseldorf</span>
+                  </a>
+                </li>
+              </ul>
+            </section>
+          </div>
+          <div className="cities">
+            <div className="cities__places-container container">
+              <section className="cities__places places">
+                <h2 className="visually-hidden">Places</h2>
+                <b className="places__found">{props.placesCount} places to stay in Amsterdam</b>
+                <form className="places__sorting" action="#" method="get">
+                  <span className="places__sorting-caption">Sort by</span>
+                  <span className="places__sorting-type" tabIndex="0">
+                    Popular
+                    <svg className="places__sorting-arrow" width="7" height="4">
+                      <use xlinkHref="#icon-arrow-select"></use>
+                    </svg>
+                  </span>
+                  <ul className="places__options places__options--custom places__options--opened">
+                    <li className="places__option places__option--active" tabIndex="0">Popular</li>
+                    <li className="places__option" tabIndex="0">Price: low to high</li>
+                    <li className="places__option" tabIndex="0">Price: high to low</li>
+                    <li className="places__option" tabIndex="0">Top rated first</li>
+                  </ul>
+                  {/* <!--
+                  <select className="places__sorting-type" id="places-sorting">
+                    <option className="places__option" value="popular" selected="">Popular</option>
+                    <option className="places__option" value="to-high">Price: low to high</option>
+                    <option className="places__option" value="to-low">Price: high to low</option>
+                    <option className="places__option" value="top-rated">Top rated first</option>
+                  </select>
+                  --> */}
+                </form>
+                <div className="cities__places-list places__list tabs__content">
+                  {props.placesNames.map((placesName, index) => {
+                    return (
+                      <article key={index} className="cities__place-card place-card">
+                        <div className="place-card__mark">
+                          <span>Premium</span>
+                        </div>
+                        <div className="cities__image-wrapper place-card__image-wrapper">
+                          <a href="#">
+                            <img className="place-card__image" src="img/apartment-01.jpg" width="260" height="200" alt="Place image" />
+                          </a>
+                        </div>
+                        <div className="place-card__info">
+                          <div className="place-card__price-wrapper">
+                            <div className="place-card__price">
+                              <b className="place-card__price-value">&euro;120</b>
+                              <span className="place-card__price-text">&#47;&nbsp;night</span>
+                            </div>
+                            <button className="place-card__bookmark-button button" type="button">
+                              <svg className="place-card__bookmark-icon" width="18" height="19">
+                                <use xlinkHref="#icon-bookmark"></use>
+                              </svg>
+                              <span className="visually-hidden">To bookmarks</span>
+                            </button>
+                          </div>
+                          <div className="place-card__rating rating">
+                            <div className="place-card__stars rating__stars">
+                              <span style={{width: `80%`}}></span>
+                              <span className="visually-hidden">Rating</span>
+                            </div>
+                          </div>
+                          <h2 className="place-card__name">
+                            <a href="#">{placesName}</a>
+                          </h2>
+                          <p className="place-card__type">Apartment</p>
+                        </div>
+                      </article>
+                    );
+                  })}
+                </div>
+              </section>
+              <div className="cities__right-section">
+                <section className="cities__map map"></section>
+              </div>
+            </div>
+          </div>
+        </main>
+      </div>
+    </React.Fragment>
+  );
+};
+
+Main.propTypes = {
+  placesCount: PropTypes.number.isRequired,
+  placesNames: PropTypes.arrayOf(PropTypes.string).isRequired
+};
+
+export default Main;
